pyproject.toml
[project]
name = 'dontrack'
version = '0.0.1'
description = 'Online donor and donation tracking'
readme = 'README.md'
requires-python = '>=3.9'
authors = [
    { name = 'promasu' },
]
classifiers = [
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
]

dependencies = [
    'Django==5.2',
    'django-auditlog==3.0.0',
    'django-bootstrap5==24.3',
    'django-qr-code==4.1.0',
    'django-localflavor==4.0',
    'gunicorn==23.0.0',
    'psycopg2-binary==2.9.10',
    'Authlib==1.5.2',
    'requests==2.32.3',
    'python-dateutil==2.9.0.post0',
]

[project.optional-dependencies]
tests = [
    'django-stubs==5.1.3',
    'mypy==1.15.0',
    'pylint==3.3.6',
    'types-requests~=2.32.0',
]
docs = [
    'mkdocs==1.6.1',
    'mkdocs-material==9.6.12'
]

[project.urls]
repository = 'https://github.com/entropia/dontrack'

[tool.setuptools.packages.find]
where=['src']

[tool.setuptools.package-data]
shiftings = ['py.typed']
